Rename clear_console_command.lua to cls_command.lua
