+++ resources/scripts/dev/commands/game_commands.lua
@@ -7,7 +7,7 @@ local food_command = require "dev.commands.food_command"
 local alignment_command = require "dev.commands.alignment_command"
 local inventory_command = require "dev.commands.inventory_command"
 local lua_command = require "dev.commands.run_lua_command"
-local clear_console_command = require "dev.commands.clear_console_command"
+local clear_console_command = require "dev.commands.cls_command"
 local condition_command = require "dev.commands.condition_command"
 local hp_command = require "dev.commands.hp_command"
 local mana_command = require "dev.commands.mana_command"
